requirements.txt
python-telegram-bot==20.5
ccxt==4.0.75
python-dotenv==1.0.0
apscheduler==3.10.4
